refactor(runtime): extract action validation into its own module
